RowIndex refactoring (#1455)
* `RowIndex` now uses primarily `size_t` in its interface, instead of `int64_t`;
* "absent" index is returned as `RowIndex::NA = size_t(-1)`, instead of `GETNA<int64_t>()`;
* min/max value is now `RowIndex::NA` for empty RowIndices, instead of 0s;
* `length()` renamed into `size()`;
* `uplift()` was replaced with `operator *`;
* `strided_loop2()` renamed into `iterate()`, and `strided_loop()` removed;
* All static RowIndex constructors were replaced with regular constructors;
* Branching for NA values during iteration is now avoided;
* `ind32` and `ind64` arrays in `ArrayRowIndexImpl` were replaced with manually managed `void*` pointer;

Closes #1440

+++ c/capi.cc
@@ -18,7 +18,7 @@ void* datatable_get_column_data(void* dt_, size_t column)
 }
 
 
-void datatable_unpack_slicerowindex(void *dt_, int64_t *start, int64_t *step)
+void datatable_unpack_slicerowindex(void* dt_, size_t* start, size_t* step)
 {
   DataTable *dt = static_cast<DataTable*>(dt_);
   RowIndex ri(dt->rowindex);
