+++ c/py_types.cc
@@ -54,45 +54,45 @@ PyObject* string_to_py(const CString& x) {
 }
 
 
-static PyObject* stype_boolean_i8_tostring(Column* col, int64_t row) {
+static PyObject* stype_boolean_i8_tostring(Column* col, size_t row) {
   int8_t x = static_cast<const int8_t*>(col->data())[row];
   return x == 0? incref(Py_False) :
          x == 1? incref(Py_True) : none();
 }
 
-static PyObject* stype_integer_i8_tostring(Column* col, int64_t row) {
+static PyObject* stype_integer_i8_tostring(Column* col, size_t row) {
   int8_t x = static_cast<const int8_t*>(col->data())[row];
   return ISNA<int8_t>(x)? none() : PyLong_FromLong(x);
 }
 
-static PyObject* stype_integer_i16_tostring(Column* col, int64_t row) {
+static PyObject* stype_integer_i16_tostring(Column* col, size_t row) {
   int16_t x = static_cast<const int16_t*>(col->data())[row];
   return ISNA<int16_t>(x)? none() : PyLong_FromLong(x);
 }
 
-static PyObject* stype_integer_i32_tostring(Column* col, int64_t row) {
+static PyObject* stype_integer_i32_tostring(Column* col, size_t row) {
   int32_t x = static_cast<const int32_t*>(col->data())[row];
   return ISNA<int32_t>(x)? none() : PyLong_FromLong(x);
 }
 
-static PyObject* stype_integer_i64_tostring(Column* col, int64_t row) {
+static PyObject* stype_integer_i64_tostring(Column* col, size_t row) {
   int64_t x = static_cast<const int64_t*>(col->data())[row];
   return ISNA<int64_t>(x)? none() : PyLong_FromLongLong(x);
 }
 
-static PyObject* stype_real_f32_tostring(Column* col, int64_t row) {
+static PyObject* stype_real_f32_tostring(Column* col, size_t row) {
   float x = static_cast<const float*>(col->data())[row];
   return ISNA<float>(x)? none() : PyFloat_FromDouble(static_cast<double>(x));
 }
 
-static PyObject* stype_real_f64_tostring(Column* col, int64_t row) {
+static PyObject* stype_real_f64_tostring(Column* col, size_t row) {
   double x = static_cast<const double*>(col->data())[row];
   return ISNA<double>(x)? none() : PyFloat_FromDouble(x);
 }
 
 
 template <typename T>
-static PyObject* stype_vchar_T_tostring(Column* col, int64_t row) {
+static PyObject* stype_vchar_T_tostring(Column* col, size_t row) {
   StringColumn<T>* str_col = static_cast<StringColumn<T>*>(col);
   const T* offsets = str_col->offsets();
   T end = offsets[row];
@@ -102,14 +102,14 @@ static PyObject* stype_vchar_T_tostring(Column* col, int64_t row) {
   return PyUnicode_FromStringAndSize(str_col->strdata() + start, len);
 }
 
-static PyObject* stype_object_pyptr_tostring(Column* col, int64_t row)
+static PyObject* stype_object_pyptr_tostring(Column* col, size_t row)
 {
   PyObject* x = static_cast<PyObject* const*>(col->data())[row];
   Py_INCREF(x);
   return x;
 }
 
-static PyObject* stype_notimpl(Column* col, int64_t)
+static PyObject* stype_notimpl(Column* col, size_t)
 {
   PyErr_Format(PyExc_NotImplementedError,
                "Cannot stringify column of type %d", col->stype());
